// frontend/src/api/models/MaaPlanConfig_Item.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MaaPlanConfig_Item = {
  /**
   * 吃理智药
   */
  MedicineNumb?: number | null
  /**
   * 连战次数
   */
  SeriesNumb?: '0' | '6' | '5' | '4' | '3' | '2' | '1' | '-1' | null
  /**
   * 关卡选择
   */
  Stage?: string | null
  /**
   * 备选关卡 - 1
   */
  Stage_1?: string | null
  /**
   * 备选关卡 - 2
   */
  Stage_2?: string | null
  /**
   * 备选关卡 - 3
   */
  Stage_3?: string | null
  /**
   * 剩余理智关卡
   */
  Stage_Remain?: string | null
}